feat(cardano-graphql): complete mapping tx fields
