+++ packages/cardano-graphql/src/Schema/types/util.ts
@@ -11,8 +11,6 @@ export type BigIntsAsStrings<T> = bigint extends T
       [K in keyof T]: T[K] extends (infer U)[] ? BigIntsAsStrings<U>[] : BigIntsAsStrings<T[K]>;
     };
 
-export type Json = string;
-
 export const Int64 = new GraphQLScalarType({
   name: 'Int64',
   parseLiteral(ast): bigint {
