+++ packages/cardano-graphql/src/sdk.ts
@@ -1575,7 +1575,7 @@ export type AddTransactionInput = {
   invalidHereafter?: Maybe<SlotRef>;
   mint?: Maybe<Array<TokenRef>>;
   outputs: Array<TransactionOutputRef>;
-  requiredExtraSignatures: Array<PublicKeyRef>;
+  requiredExtraSignatures?: Maybe<Array<PublicKeyRef>>;
   scriptIntegrityHash?: Maybe<Scalars['String']>;
   size: Scalars['Int64'];
   totalOutputCoin: Scalars['Int64'];
@@ -9731,7 +9731,7 @@ export type Transaction = {
   mintAggregate?: Maybe<TokenAggregateResult>;
   outputs: Array<TransactionOutput>;
   outputsAggregate?: Maybe<TransactionOutputAggregateResult>;
-  requiredExtraSignatures: Array<PublicKey>;
+  requiredExtraSignatures?: Maybe<Array<PublicKey>>;
   requiredExtraSignaturesAggregate?: Maybe<PublicKeyAggregateResult>;
   scriptIntegrityHash?: Maybe<Scalars['String']>;
   size: Scalars['Int64'];
@@ -11862,21 +11862,21 @@ type AnyScript_PlutusScript_Fragment = { __typename: 'PlutusScript', cborHex: st
 
 export type AnyScriptFragment = AnyScript_NativeScript_Fragment | AnyScript_PlutusScript_Fragment;
 
-export type CoreTransactionFieldsFragment = { __typename?: 'Transaction', fee: bigint, hash: string, index: number, size: bigint, inputs: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }>, outputs: Array<{ __typename?: 'TransactionOutput', datumHash?: string | null | undefined, address: { __typename?: 'Address', address: string }, value: { __typename?: 'Value', coin: bigint, assets?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined } }>, certificates?: Array<{ __typename: 'GenesisKeyDelegationCertificate', genesisHash: string, genesisDelegateHash: string, vrfKeyHash: string } | { __typename: 'MirCertificate', quantity: bigint, pot: string, rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'PoolRegistrationCertificate', epoch: { __typename?: 'Epoch', number: number }, poolParameters: { __typename?: 'PoolParameters', cost: bigint, vrf: string, pledge: bigint, stakePool: { __typename?: 'StakePool', id: string }, metadata?: { __typename?: 'StakePoolMetadata', ticker: string, name: string, description: string, homepage: string, extDataUrl?: string | null | undefined, extSigUrl?: string | null | undefined, extVkey?: string | null | undefined, ext?: { __typename?: 'ExtendedStakePoolMetadata', serial: number, pool: { __typename?: 'ExtendedStakePoolMetadataFields', id: string, country?: string | null | undefined, status?: ExtendedPoolStatus | null | undefined, contact?: { __typename?: 'PoolContactData', primary: string, email?: string | null | undefined, facebook?: string | null | undefined, github?: string | null | undefined, feed?: string | null | undefined, telegram?: string | null | undefined, twitter?: string | null | undefined } | null | undefined, media_assets?: { __typename?: 'ThePoolsMediaAssets', icon_png_64x64: string, logo_png?: string | null | undefined, logo_svg?: string | null | undefined, color_fg?: string | null | undefined, color_bg?: string | null | undefined } | null | undefined, itn?: { __typename?: 'ITNVerification', owner: string, witness: string } | null | undefined } } | null | undefined } | null | undefined, owners: Array<{ __typename?: 'RewardAccount', address: string }>, margin: { __typename?: 'Ratio', numerator: number, denominator: number }, relays: Array<{ __typename: 'RelayByAddress', ipv4?: string | null | undefined, ipv6?: string | null | undefined, port?: number | null | undefined } | { __typename: 'RelayByName', hostname: string, port?: number | null | undefined } | { __typename: 'RelayByNameMultihost', dnsName: string }>, poolRegistrationCertificate: { __typename?: 'PoolRegistrationCertificate', transaction: { __typename?: 'Transaction', hash: string } }, rewardAccount: { __typename?: 'RewardAccount', address: string }, metadataJson?: { __typename?: 'StakePoolMetadataJson', hash: string, url: string } | null | undefined } } | { __typename: 'PoolRetirementCertificate', epoch: { __typename?: 'Epoch', number: number }, stakePool: { __typename?: 'StakePool', id: string } } | { __typename: 'StakeDelegationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string }, stakePool: { __typename?: 'StakePool', id: string }, epoch: { __typename?: 'Epoch', number: number } } | { __typename: 'StakeKeyDeregistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'StakeKeyRegistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, collateral?: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }> | null | undefined, invalidBefore?: { __typename?: 'Slot', slotNo: number } | null | undefined, invalidHereafter?: { __typename?: 'Slot', slotNo: number } | null | undefined, withdrawals?: Array<{ __typename?: 'Withdrawal', quantity: bigint, rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, mint?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined, block: { __typename?: 'Block', blockNo: number, hash: string, slot: { __typename?: 'Slot', number: number } }, witness: { __typename?: 'Witness', signatures: Array<{ __typename?: 'Signature', signature: string, publicKey: { __typename?: 'PublicKey', key: string } }>, scripts?: Array<{ __typename?: 'WitnessScript', key: string, script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, bootstrap?: Array<{ __typename?: 'BootstrapWitness', signature: string, chainCode?: string | null | undefined, addressAttributes?: string | null | undefined, key?: { __typename?: 'PublicKey', hash: string, key: string } | null | undefined }> | null | undefined, redeemers?: Array<{ __typename?: 'Redeemer', index: number, purpose: string, scriptHash: string, executionUnits: { __typename?: 'ExecutionUnits', memory: number, steps: number } }> | null | undefined, datums?: Array<{ __typename?: 'Datum', hash: string, datum: string }> | null | undefined }, auxiliaryData?: { __typename?: 'AuxiliaryData', hash: string, body: { __typename?: 'AuxiliaryDataBody', blob?: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> | null | undefined } } | null | undefined };
+export type CoreTransactionFieldsFragment = { __typename?: 'Transaction', fee: bigint, hash: string, index: number, size: bigint, scriptIntegrityHash?: string | null | undefined, inputs: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }>, outputs: Array<{ __typename?: 'TransactionOutput', datumHash?: string | null | undefined, address: { __typename?: 'Address', address: string }, value: { __typename?: 'Value', coin: bigint, assets?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined } }>, certificates?: Array<{ __typename: 'GenesisKeyDelegationCertificate', genesisHash: string, genesisDelegateHash: string, vrfKeyHash: string } | { __typename: 'MirCertificate', quantity: bigint, pot: string, rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'PoolRegistrationCertificate', epoch: { __typename?: 'Epoch', number: number }, poolParameters: { __typename?: 'PoolParameters', cost: bigint, vrf: string, pledge: bigint, stakePool: { __typename?: 'StakePool', id: string }, metadata?: { __typename?: 'StakePoolMetadata', ticker: string, name: string, description: string, homepage: string, extDataUrl?: string | null | undefined, extSigUrl?: string | null | undefined, extVkey?: string | null | undefined, ext?: { __typename?: 'ExtendedStakePoolMetadata', serial: number, pool: { __typename?: 'ExtendedStakePoolMetadataFields', id: string, country?: string | null | undefined, status?: ExtendedPoolStatus | null | undefined, contact?: { __typename?: 'PoolContactData', primary: string, email?: string | null | undefined, facebook?: string | null | undefined, github?: string | null | undefined, feed?: string | null | undefined, telegram?: string | null | undefined, twitter?: string | null | undefined } | null | undefined, media_assets?: { __typename?: 'ThePoolsMediaAssets', icon_png_64x64: string, logo_png?: string | null | undefined, logo_svg?: string | null | undefined, color_fg?: string | null | undefined, color_bg?: string | null | undefined } | null | undefined, itn?: { __typename?: 'ITNVerification', owner: string, witness: string } | null | undefined } } | null | undefined } | null | undefined, owners: Array<{ __typename?: 'RewardAccount', address: string }>, margin: { __typename?: 'Ratio', numerator: number, denominator: number }, relays: Array<{ __typename: 'RelayByAddress', ipv4?: string | null | undefined, ipv6?: string | null | undefined, port?: number | null | undefined } | { __typename: 'RelayByName', hostname: string, port?: number | null | undefined } | { __typename: 'RelayByNameMultihost', dnsName: string }>, poolRegistrationCertificate: { __typename?: 'PoolRegistrationCertificate', transaction: { __typename?: 'Transaction', hash: string } }, rewardAccount: { __typename?: 'RewardAccount', address: string }, metadataJson?: { __typename?: 'StakePoolMetadataJson', hash: string, url: string } | null | undefined } } | { __typename: 'PoolRetirementCertificate', epoch: { __typename?: 'Epoch', number: number }, stakePool: { __typename?: 'StakePool', id: string } } | { __typename: 'StakeDelegationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string }, stakePool: { __typename?: 'StakePool', id: string }, epoch: { __typename?: 'Epoch', number: number } } | { __typename: 'StakeKeyDeregistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'StakeKeyRegistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, collateral?: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }> | null | undefined, invalidBefore?: { __typename?: 'Slot', slotNo: number } | null | undefined, invalidHereafter?: { __typename?: 'Slot', slotNo: number } | null | undefined, withdrawals?: Array<{ __typename?: 'Withdrawal', quantity: bigint, rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, mint?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined, block: { __typename?: 'Block', blockNo: number, hash: string, slot: { __typename?: 'Slot', number: number } }, requiredExtraSignatures?: Array<{ __typename?: 'PublicKey', hash: string }> | null | undefined, witness: { __typename?: 'Witness', signatures: Array<{ __typename?: 'Signature', signature: string, publicKey: { __typename?: 'PublicKey', key: string } }>, scripts?: Array<{ __typename?: 'WitnessScript', key: string, script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, bootstrap?: Array<{ __typename?: 'BootstrapWitness', signature: string, chainCode?: string | null | undefined, addressAttributes?: string | null | undefined, key?: { __typename?: 'PublicKey', hash: string, key: string } | null | undefined }> | null | undefined, redeemers?: Array<{ __typename?: 'Redeemer', index: number, purpose: string, scriptHash: string, executionUnits: { __typename?: 'ExecutionUnits', memory: number, steps: number } }> | null | undefined, datums?: Array<{ __typename?: 'Datum', hash: string, datum: string }> | null | undefined }, auxiliaryData?: { __typename?: 'AuxiliaryData', hash: string, body: { __typename?: 'AuxiliaryDataBody', scripts?: Array<{ __typename?: 'AuxiliaryScript', script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, blob?: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> | null | undefined } } | null | undefined };
 
 export type TransactionsByHashesQueryVariables = Exact<{
   hashes: Array<Scalars['String']> | Scalars['String'];
 }>;
 
 
-export type TransactionsByHashesQuery = { __typename?: 'Query', queryProtocolParametersAlonzo?: Array<{ __typename?: 'ProtocolParametersAlonzo', stakeKeyDeposit: number, poolDeposit: number } | null | undefined> | null | undefined, queryTransaction?: Array<{ __typename?: 'Transaction', fee: bigint, hash: string, index: number, size: bigint, inputs: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }>, outputs: Array<{ __typename?: 'TransactionOutput', datumHash?: string | null | undefined, address: { __typename?: 'Address', address: string }, value: { __typename?: 'Value', coin: bigint, assets?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined } }>, certificates?: Array<{ __typename: 'GenesisKeyDelegationCertificate', genesisHash: string, genesisDelegateHash: string, vrfKeyHash: string } | { __typename: 'MirCertificate', quantity: bigint, pot: string, rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'PoolRegistrationCertificate', epoch: { __typename?: 'Epoch', number: number }, poolParameters: { __typename?: 'PoolParameters', cost: bigint, vrf: string, pledge: bigint, stakePool: { __typename?: 'StakePool', id: string }, metadata?: { __typename?: 'StakePoolMetadata', ticker: string, name: string, description: string, homepage: string, extDataUrl?: string | null | undefined, extSigUrl?: string | null | undefined, extVkey?: string | null | undefined, ext?: { __typename?: 'ExtendedStakePoolMetadata', serial: number, pool: { __typename?: 'ExtendedStakePoolMetadataFields', id: string, country?: string | null | undefined, status?: ExtendedPoolStatus | null | undefined, contact?: { __typename?: 'PoolContactData', primary: string, email?: string | null | undefined, facebook?: string | null | undefined, github?: string | null | undefined, feed?: string | null | undefined, telegram?: string | null | undefined, twitter?: string | null | undefined } | null | undefined, media_assets?: { __typename?: 'ThePoolsMediaAssets', icon_png_64x64: string, logo_png?: string | null | undefined, logo_svg?: string | null | undefined, color_fg?: string | null | undefined, color_bg?: string | null | undefined } | null | undefined, itn?: { __typename?: 'ITNVerification', owner: string, witness: string } | null | undefined } } | null | undefined } | null | undefined, owners: Array<{ __typename?: 'RewardAccount', address: string }>, margin: { __typename?: 'Ratio', numerator: number, denominator: number }, relays: Array<{ __typename: 'RelayByAddress', ipv4?: string | null | undefined, ipv6?: string | null | undefined, port?: number | null | undefined } | { __typename: 'RelayByName', hostname: string, port?: number | null | undefined } | { __typename: 'RelayByNameMultihost', dnsName: string }>, poolRegistrationCertificate: { __typename?: 'PoolRegistrationCertificate', transaction: { __typename?: 'Transaction', hash: string } }, rewardAccount: { __typename?: 'RewardAccount', address: string }, metadataJson?: { __typename?: 'StakePoolMetadataJson', hash: string, url: string } | null | undefined } } | { __typename: 'PoolRetirementCertificate', epoch: { __typename?: 'Epoch', number: number }, stakePool: { __typename?: 'StakePool', id: string } } | { __typename: 'StakeDelegationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string }, stakePool: { __typename?: 'StakePool', id: string }, epoch: { __typename?: 'Epoch', number: number } } | { __typename: 'StakeKeyDeregistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'StakeKeyRegistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, collateral?: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }> | null | undefined, invalidBefore?: { __typename?: 'Slot', slotNo: number } | null | undefined, invalidHereafter?: { __typename?: 'Slot', slotNo: number } | null | undefined, withdrawals?: Array<{ __typename?: 'Withdrawal', quantity: bigint, rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, mint?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined, block: { __typename?: 'Block', blockNo: number, hash: string, slot: { __typename?: 'Slot', number: number } }, witness: { __typename?: 'Witness', signatures: Array<{ __typename?: 'Signature', signature: string, publicKey: { __typename?: 'PublicKey', key: string } }>, scripts?: Array<{ __typename?: 'WitnessScript', key: string, script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, bootstrap?: Array<{ __typename?: 'BootstrapWitness', signature: string, chainCode?: string | null | undefined, addressAttributes?: string | null | undefined, key?: { __typename?: 'PublicKey', hash: string, key: string } | null | undefined }> | null | undefined, redeemers?: Array<{ __typename?: 'Redeemer', index: number, purpose: string, scriptHash: string, executionUnits: { __typename?: 'ExecutionUnits', memory: number, steps: number } }> | null | undefined, datums?: Array<{ __typename?: 'Datum', hash: string, datum: string }> | null | undefined }, auxiliaryData?: { __typename?: 'AuxiliaryData', hash: string, body: { __typename?: 'AuxiliaryDataBody', blob?: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> | null | undefined } } | null | undefined } | null | undefined> | null | undefined };
+export type TransactionsByHashesQuery = { __typename?: 'Query', queryProtocolParametersAlonzo?: Array<{ __typename?: 'ProtocolParametersAlonzo', stakeKeyDeposit: number, poolDeposit: number } | null | undefined> | null | undefined, queryTransaction?: Array<{ __typename?: 'Transaction', fee: bigint, hash: string, index: number, size: bigint, scriptIntegrityHash?: string | null | undefined, inputs: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }>, outputs: Array<{ __typename?: 'TransactionOutput', datumHash?: string | null | undefined, address: { __typename?: 'Address', address: string }, value: { __typename?: 'Value', coin: bigint, assets?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined } }>, certificates?: Array<{ __typename: 'GenesisKeyDelegationCertificate', genesisHash: string, genesisDelegateHash: string, vrfKeyHash: string } | { __typename: 'MirCertificate', quantity: bigint, pot: string, rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'PoolRegistrationCertificate', epoch: { __typename?: 'Epoch', number: number }, poolParameters: { __typename?: 'PoolParameters', cost: bigint, vrf: string, pledge: bigint, stakePool: { __typename?: 'StakePool', id: string }, metadata?: { __typename?: 'StakePoolMetadata', ticker: string, name: string, description: string, homepage: string, extDataUrl?: string | null | undefined, extSigUrl?: string | null | undefined, extVkey?: string | null | undefined, ext?: { __typename?: 'ExtendedStakePoolMetadata', serial: number, pool: { __typename?: 'ExtendedStakePoolMetadataFields', id: string, country?: string | null | undefined, status?: ExtendedPoolStatus | null | undefined, contact?: { __typename?: 'PoolContactData', primary: string, email?: string | null | undefined, facebook?: string | null | undefined, github?: string | null | undefined, feed?: string | null | undefined, telegram?: string | null | undefined, twitter?: string | null | undefined } | null | undefined, media_assets?: { __typename?: 'ThePoolsMediaAssets', icon_png_64x64: string, logo_png?: string | null | undefined, logo_svg?: string | null | undefined, color_fg?: string | null | undefined, color_bg?: string | null | undefined } | null | undefined, itn?: { __typename?: 'ITNVerification', owner: string, witness: string } | null | undefined } } | null | undefined } | null | undefined, owners: Array<{ __typename?: 'RewardAccount', address: string }>, margin: { __typename?: 'Ratio', numerator: number, denominator: number }, relays: Array<{ __typename: 'RelayByAddress', ipv4?: string | null | undefined, ipv6?: string | null | undefined, port?: number | null | undefined } | { __typename: 'RelayByName', hostname: string, port?: number | null | undefined } | { __typename: 'RelayByNameMultihost', dnsName: string }>, poolRegistrationCertificate: { __typename?: 'PoolRegistrationCertificate', transaction: { __typename?: 'Transaction', hash: string } }, rewardAccount: { __typename?: 'RewardAccount', address: string }, metadataJson?: { __typename?: 'StakePoolMetadataJson', hash: string, url: string } | null | undefined } } | { __typename: 'PoolRetirementCertificate', epoch: { __typename?: 'Epoch', number: number }, stakePool: { __typename?: 'StakePool', id: string } } | { __typename: 'StakeDelegationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string }, stakePool: { __typename?: 'StakePool', id: string }, epoch: { __typename?: 'Epoch', number: number } } | { __typename: 'StakeKeyDeregistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'StakeKeyRegistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, collateral?: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }> | null | undefined, invalidBefore?: { __typename?: 'Slot', slotNo: number } | null | undefined, invalidHereafter?: { __typename?: 'Slot', slotNo: number } | null | undefined, withdrawals?: Array<{ __typename?: 'Withdrawal', quantity: bigint, rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, mint?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined, block: { __typename?: 'Block', blockNo: number, hash: string, slot: { __typename?: 'Slot', number: number } }, requiredExtraSignatures?: Array<{ __typename?: 'PublicKey', hash: string }> | null | undefined, witness: { __typename?: 'Witness', signatures: Array<{ __typename?: 'Signature', signature: string, publicKey: { __typename?: 'PublicKey', key: string } }>, scripts?: Array<{ __typename?: 'WitnessScript', key: string, script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, bootstrap?: Array<{ __typename?: 'BootstrapWitness', signature: string, chainCode?: string | null | undefined, addressAttributes?: string | null | undefined, key?: { __typename?: 'PublicKey', hash: string, key: string } | null | undefined }> | null | undefined, redeemers?: Array<{ __typename?: 'Redeemer', index: number, purpose: string, scriptHash: string, executionUnits: { __typename?: 'ExecutionUnits', memory: number, steps: number } }> | null | undefined, datums?: Array<{ __typename?: 'Datum', hash: string, datum: string }> | null | undefined }, auxiliaryData?: { __typename?: 'AuxiliaryData', hash: string, body: { __typename?: 'AuxiliaryDataBody', scripts?: Array<{ __typename?: 'AuxiliaryScript', script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, blob?: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> | null | undefined } } | null | undefined } | null | undefined> | null | undefined };
 
 export type TransactionIdsByAddressesQueryVariables = Exact<{
   addresses: Array<Scalars['String']> | Scalars['String'];
 }>;
 
 
-export type TransactionIdsByAddressesQuery = { __typename?: 'Query', queryProtocolParametersAlonzo?: Array<{ __typename?: 'ProtocolParametersAlonzo', stakeKeyDeposit: number, poolDeposit: number } | null | undefined> | null | undefined, queryAddress?: Array<{ __typename?: 'Address', inputs: Array<{ __typename?: 'TransactionInput', transaction: { __typename?: 'Transaction', fee: bigint, hash: string, index: number, size: bigint, inputs: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }>, outputs: Array<{ __typename?: 'TransactionOutput', datumHash?: string | null | undefined, address: { __typename?: 'Address', address: string }, value: { __typename?: 'Value', coin: bigint, assets?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined } }>, certificates?: Array<{ __typename: 'GenesisKeyDelegationCertificate', genesisHash: string, genesisDelegateHash: string, vrfKeyHash: string } | { __typename: 'MirCertificate', quantity: bigint, pot: string, rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'PoolRegistrationCertificate', epoch: { __typename?: 'Epoch', number: number }, poolParameters: { __typename?: 'PoolParameters', cost: bigint, vrf: string, pledge: bigint, stakePool: { __typename?: 'StakePool', id: string }, metadata?: { __typename?: 'StakePoolMetadata', ticker: string, name: string, description: string, homepage: string, extDataUrl?: string | null | undefined, extSigUrl?: string | null | undefined, extVkey?: string | null | undefined, ext?: { __typename?: 'ExtendedStakePoolMetadata', serial: number, pool: { __typename?: 'ExtendedStakePoolMetadataFields', id: string, country?: string | null | undefined, status?: ExtendedPoolStatus | null | undefined, contact?: { __typename?: 'PoolContactData', primary: string, email?: string | null | undefined, facebook?: string | null | undefined, github?: string | null | undefined, feed?: string | null | undefined, telegram?: string | null | undefined, twitter?: string | null | undefined } | null | undefined, media_assets?: { __typename?: 'ThePoolsMediaAssets', icon_png_64x64: string, logo_png?: string | null | undefined, logo_svg?: string | null | undefined, color_fg?: string | null | undefined, color_bg?: string | null | undefined } | null | undefined, itn?: { __typename?: 'ITNVerification', owner: string, witness: string } | null | undefined } } | null | undefined } | null | undefined, owners: Array<{ __typename?: 'RewardAccount', address: string }>, margin: { __typename?: 'Ratio', numerator: number, denominator: number }, relays: Array<{ __typename: 'RelayByAddress', ipv4?: string | null | undefined, ipv6?: string | null | undefined, port?: number | null | undefined } | { __typename: 'RelayByName', hostname: string, port?: number | null | undefined } | { __typename: 'RelayByNameMultihost', dnsName: string }>, poolRegistrationCertificate: { __typename?: 'PoolRegistrationCertificate', transaction: { __typename?: 'Transaction', hash: string } }, rewardAccount: { __typename?: 'RewardAccount', address: string }, metadataJson?: { __typename?: 'StakePoolMetadataJson', hash: string, url: string } | null | undefined } } | { __typename: 'PoolRetirementCertificate', epoch: { __typename?: 'Epoch', number: number }, stakePool: { __typename?: 'StakePool', id: string } } | { __typename: 'StakeDelegationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string }, stakePool: { __typename?: 'StakePool', id: string }, epoch: { __typename?: 'Epoch', number: number } } | { __typename: 'StakeKeyDeregistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'StakeKeyRegistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, collateral?: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }> | null | undefined, invalidBefore?: { __typename?: 'Slot', slotNo: number } | null | undefined, invalidHereafter?: { __typename?: 'Slot', slotNo: number } | null | undefined, withdrawals?: Array<{ __typename?: 'Withdrawal', quantity: bigint, rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, mint?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined, block: { __typename?: 'Block', blockNo: number, hash: string, slot: { __typename?: 'Slot', number: number } }, witness: { __typename?: 'Witness', signatures: Array<{ __typename?: 'Signature', signature: string, publicKey: { __typename?: 'PublicKey', key: string } }>, scripts?: Array<{ __typename?: 'WitnessScript', key: string, script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, bootstrap?: Array<{ __typename?: 'BootstrapWitness', signature: string, chainCode?: string | null | undefined, addressAttributes?: string | null | undefined, key?: { __typename?: 'PublicKey', hash: string, key: string } | null | undefined }> | null | undefined, redeemers?: Array<{ __typename?: 'Redeemer', index: number, purpose: string, scriptHash: string, executionUnits: { __typename?: 'ExecutionUnits', memory: number, steps: number } }> | null | undefined, datums?: Array<{ __typename?: 'Datum', hash: string, datum: string }> | null | undefined }, auxiliaryData?: { __typename?: 'AuxiliaryData', hash: string, body: { __typename?: 'AuxiliaryDataBody', blob?: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> | null | undefined } } | null | undefined } }>, utxo: Array<{ __typename?: 'TransactionOutput', transaction: { __typename?: 'Transaction', fee: bigint, hash: string, index: number, size: bigint, inputs: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }>, outputs: Array<{ __typename?: 'TransactionOutput', datumHash?: string | null | undefined, address: { __typename?: 'Address', address: string }, value: { __typename?: 'Value', coin: bigint, assets?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined } }>, certificates?: Array<{ __typename: 'GenesisKeyDelegationCertificate', genesisHash: string, genesisDelegateHash: string, vrfKeyHash: string } | { __typename: 'MirCertificate', quantity: bigint, pot: string, rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'PoolRegistrationCertificate', epoch: { __typename?: 'Epoch', number: number }, poolParameters: { __typename?: 'PoolParameters', cost: bigint, vrf: string, pledge: bigint, stakePool: { __typename?: 'StakePool', id: string }, metadata?: { __typename?: 'StakePoolMetadata', ticker: string, name: string, description: string, homepage: string, extDataUrl?: string | null | undefined, extSigUrl?: string | null | undefined, extVkey?: string | null | undefined, ext?: { __typename?: 'ExtendedStakePoolMetadata', serial: number, pool: { __typename?: 'ExtendedStakePoolMetadataFields', id: string, country?: string | null | undefined, status?: ExtendedPoolStatus | null | undefined, contact?: { __typename?: 'PoolContactData', primary: string, email?: string | null | undefined, facebook?: string | null | undefined, github?: string | null | undefined, feed?: string | null | undefined, telegram?: string | null | undefined, twitter?: string | null | undefined } | null | undefined, media_assets?: { __typename?: 'ThePoolsMediaAssets', icon_png_64x64: string, logo_png?: string | null | undefined, logo_svg?: string | null | undefined, color_fg?: string | null | undefined, color_bg?: string | null | undefined } | null | undefined, itn?: { __typename?: 'ITNVerification', owner: string, witness: string } | null | undefined } } | null | undefined } | null | undefined, owners: Array<{ __typename?: 'RewardAccount', address: string }>, margin: { __typename?: 'Ratio', numerator: number, denominator: number }, relays: Array<{ __typename: 'RelayByAddress', ipv4?: string | null | undefined, ipv6?: string | null | undefined, port?: number | null | undefined } | { __typename: 'RelayByName', hostname: string, port?: number | null | undefined } | { __typename: 'RelayByNameMultihost', dnsName: string }>, poolRegistrationCertificate: { __typename?: 'PoolRegistrationCertificate', transaction: { __typename?: 'Transaction', hash: string } }, rewardAccount: { __typename?: 'RewardAccount', address: string }, metadataJson?: { __typename?: 'StakePoolMetadataJson', hash: string, url: string } | null | undefined } } | { __typename: 'PoolRetirementCertificate', epoch: { __typename?: 'Epoch', number: number }, stakePool: { __typename?: 'StakePool', id: string } } | { __typename: 'StakeDelegationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string }, stakePool: { __typename?: 'StakePool', id: string }, epoch: { __typename?: 'Epoch', number: number } } | { __typename: 'StakeKeyDeregistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'StakeKeyRegistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, collateral?: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }> | null | undefined, invalidBefore?: { __typename?: 'Slot', slotNo: number } | null | undefined, invalidHereafter?: { __typename?: 'Slot', slotNo: number } | null | undefined, withdrawals?: Array<{ __typename?: 'Withdrawal', quantity: bigint, rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, mint?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined, block: { __typename?: 'Block', blockNo: number, hash: string, slot: { __typename?: 'Slot', number: number } }, witness: { __typename?: 'Witness', signatures: Array<{ __typename?: 'Signature', signature: string, publicKey: { __typename?: 'PublicKey', key: string } }>, scripts?: Array<{ __typename?: 'WitnessScript', key: string, script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, bootstrap?: Array<{ __typename?: 'BootstrapWitness', signature: string, chainCode?: string | null | undefined, addressAttributes?: string | null | undefined, key?: { __typename?: 'PublicKey', hash: string, key: string } | null | undefined }> | null | undefined, redeemers?: Array<{ __typename?: 'Redeemer', index: number, purpose: string, scriptHash: string, executionUnits: { __typename?: 'ExecutionUnits', memory: number, steps: number } }> | null | undefined, datums?: Array<{ __typename?: 'Datum', hash: string, datum: string }> | null | undefined }, auxiliaryData?: { __typename?: 'AuxiliaryData', hash: string, body: { __typename?: 'AuxiliaryDataBody', blob?: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> | null | undefined } } | null | undefined } }> } | null | undefined> | null | undefined };
+export type TransactionIdsByAddressesQuery = { __typename?: 'Query', queryProtocolParametersAlonzo?: Array<{ __typename?: 'ProtocolParametersAlonzo', stakeKeyDeposit: number, poolDeposit: number } | null | undefined> | null | undefined, queryAddress?: Array<{ __typename?: 'Address', inputs: Array<{ __typename?: 'TransactionInput', transaction: { __typename?: 'Transaction', fee: bigint, hash: string, index: number, size: bigint, scriptIntegrityHash?: string | null | undefined, inputs: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }>, outputs: Array<{ __typename?: 'TransactionOutput', datumHash?: string | null | undefined, address: { __typename?: 'Address', address: string }, value: { __typename?: 'Value', coin: bigint, assets?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined } }>, certificates?: Array<{ __typename: 'GenesisKeyDelegationCertificate', genesisHash: string, genesisDelegateHash: string, vrfKeyHash: string } | { __typename: 'MirCertificate', quantity: bigint, pot: string, rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'PoolRegistrationCertificate', epoch: { __typename?: 'Epoch', number: number }, poolParameters: { __typename?: 'PoolParameters', cost: bigint, vrf: string, pledge: bigint, stakePool: { __typename?: 'StakePool', id: string }, metadata?: { __typename?: 'StakePoolMetadata', ticker: string, name: string, description: string, homepage: string, extDataUrl?: string | null | undefined, extSigUrl?: string | null | undefined, extVkey?: string | null | undefined, ext?: { __typename?: 'ExtendedStakePoolMetadata', serial: number, pool: { __typename?: 'ExtendedStakePoolMetadataFields', id: string, country?: string | null | undefined, status?: ExtendedPoolStatus | null | undefined, contact?: { __typename?: 'PoolContactData', primary: string, email?: string | null | undefined, facebook?: string | null | undefined, github?: string | null | undefined, feed?: string | null | undefined, telegram?: string | null | undefined, twitter?: string | null | undefined } | null | undefined, media_assets?: { __typename?: 'ThePoolsMediaAssets', icon_png_64x64: string, logo_png?: string | null | undefined, logo_svg?: string | null | undefined, color_fg?: string | null | undefined, color_bg?: string | null | undefined } | null | undefined, itn?: { __typename?: 'ITNVerification', owner: string, witness: string } | null | undefined } } | null | undefined } | null | undefined, owners: Array<{ __typename?: 'RewardAccount', address: string }>, margin: { __typename?: 'Ratio', numerator: number, denominator: number }, relays: Array<{ __typename: 'RelayByAddress', ipv4?: string | null | undefined, ipv6?: string | null | undefined, port?: number | null | undefined } | { __typename: 'RelayByName', hostname: string, port?: number | null | undefined } | { __typename: 'RelayByNameMultihost', dnsName: string }>, poolRegistrationCertificate: { __typename?: 'PoolRegistrationCertificate', transaction: { __typename?: 'Transaction', hash: string } }, rewardAccount: { __typename?: 'RewardAccount', address: string }, metadataJson?: { __typename?: 'StakePoolMetadataJson', hash: string, url: string } | null | undefined } } | { __typename: 'PoolRetirementCertificate', epoch: { __typename?: 'Epoch', number: number }, stakePool: { __typename?: 'StakePool', id: string } } | { __typename: 'StakeDelegationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string }, stakePool: { __typename?: 'StakePool', id: string }, epoch: { __typename?: 'Epoch', number: number } } | { __typename: 'StakeKeyDeregistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'StakeKeyRegistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, collateral?: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }> | null | undefined, invalidBefore?: { __typename?: 'Slot', slotNo: number } | null | undefined, invalidHereafter?: { __typename?: 'Slot', slotNo: number } | null | undefined, withdrawals?: Array<{ __typename?: 'Withdrawal', quantity: bigint, rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, mint?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined, block: { __typename?: 'Block', blockNo: number, hash: string, slot: { __typename?: 'Slot', number: number } }, requiredExtraSignatures?: Array<{ __typename?: 'PublicKey', hash: string }> | null | undefined, witness: { __typename?: 'Witness', signatures: Array<{ __typename?: 'Signature', signature: string, publicKey: { __typename?: 'PublicKey', key: string } }>, scripts?: Array<{ __typename?: 'WitnessScript', key: string, script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, bootstrap?: Array<{ __typename?: 'BootstrapWitness', signature: string, chainCode?: string | null | undefined, addressAttributes?: string | null | undefined, key?: { __typename?: 'PublicKey', hash: string, key: string } | null | undefined }> | null | undefined, redeemers?: Array<{ __typename?: 'Redeemer', index: number, purpose: string, scriptHash: string, executionUnits: { __typename?: 'ExecutionUnits', memory: number, steps: number } }> | null | undefined, datums?: Array<{ __typename?: 'Datum', hash: string, datum: string }> | null | undefined }, auxiliaryData?: { __typename?: 'AuxiliaryData', hash: string, body: { __typename?: 'AuxiliaryDataBody', scripts?: Array<{ __typename?: 'AuxiliaryScript', script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, blob?: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> | null | undefined } } | null | undefined } }>, utxo: Array<{ __typename?: 'TransactionOutput', transaction: { __typename?: 'Transaction', fee: bigint, hash: string, index: number, size: bigint, scriptIntegrityHash?: string | null | undefined, inputs: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }>, outputs: Array<{ __typename?: 'TransactionOutput', datumHash?: string | null | undefined, address: { __typename?: 'Address', address: string }, value: { __typename?: 'Value', coin: bigint, assets?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined } }>, certificates?: Array<{ __typename: 'GenesisKeyDelegationCertificate', genesisHash: string, genesisDelegateHash: string, vrfKeyHash: string } | { __typename: 'MirCertificate', quantity: bigint, pot: string, rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'PoolRegistrationCertificate', epoch: { __typename?: 'Epoch', number: number }, poolParameters: { __typename?: 'PoolParameters', cost: bigint, vrf: string, pledge: bigint, stakePool: { __typename?: 'StakePool', id: string }, metadata?: { __typename?: 'StakePoolMetadata', ticker: string, name: string, description: string, homepage: string, extDataUrl?: string | null | undefined, extSigUrl?: string | null | undefined, extVkey?: string | null | undefined, ext?: { __typename?: 'ExtendedStakePoolMetadata', serial: number, pool: { __typename?: 'ExtendedStakePoolMetadataFields', id: string, country?: string | null | undefined, status?: ExtendedPoolStatus | null | undefined, contact?: { __typename?: 'PoolContactData', primary: string, email?: string | null | undefined, facebook?: string | null | undefined, github?: string | null | undefined, feed?: string | null | undefined, telegram?: string | null | undefined, twitter?: string | null | undefined } | null | undefined, media_assets?: { __typename?: 'ThePoolsMediaAssets', icon_png_64x64: string, logo_png?: string | null | undefined, logo_svg?: string | null | undefined, color_fg?: string | null | undefined, color_bg?: string | null | undefined } | null | undefined, itn?: { __typename?: 'ITNVerification', owner: string, witness: string } | null | undefined } } | null | undefined } | null | undefined, owners: Array<{ __typename?: 'RewardAccount', address: string }>, margin: { __typename?: 'Ratio', numerator: number, denominator: number }, relays: Array<{ __typename: 'RelayByAddress', ipv4?: string | null | undefined, ipv6?: string | null | undefined, port?: number | null | undefined } | { __typename: 'RelayByName', hostname: string, port?: number | null | undefined } | { __typename: 'RelayByNameMultihost', dnsName: string }>, poolRegistrationCertificate: { __typename?: 'PoolRegistrationCertificate', transaction: { __typename?: 'Transaction', hash: string } }, rewardAccount: { __typename?: 'RewardAccount', address: string }, metadataJson?: { __typename?: 'StakePoolMetadataJson', hash: string, url: string } | null | undefined } } | { __typename: 'PoolRetirementCertificate', epoch: { __typename?: 'Epoch', number: number }, stakePool: { __typename?: 'StakePool', id: string } } | { __typename: 'StakeDelegationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string }, stakePool: { __typename?: 'StakePool', id: string }, epoch: { __typename?: 'Epoch', number: number } } | { __typename: 'StakeKeyDeregistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } } | { __typename: 'StakeKeyRegistrationCertificate', rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, collateral?: Array<{ __typename?: 'TransactionInput', index: number, address: { __typename?: 'Address', address: string } }> | null | undefined, invalidBefore?: { __typename?: 'Slot', slotNo: number } | null | undefined, invalidHereafter?: { __typename?: 'Slot', slotNo: number } | null | undefined, withdrawals?: Array<{ __typename?: 'Withdrawal', quantity: bigint, rewardAccount: { __typename?: 'RewardAccount', address: string } }> | null | undefined, mint?: Array<{ __typename?: 'Token', quantity: string, asset: { __typename?: 'Asset', assetId: string } }> | null | undefined, block: { __typename?: 'Block', blockNo: number, hash: string, slot: { __typename?: 'Slot', number: number } }, requiredExtraSignatures?: Array<{ __typename?: 'PublicKey', hash: string }> | null | undefined, witness: { __typename?: 'Witness', signatures: Array<{ __typename?: 'Signature', signature: string, publicKey: { __typename?: 'PublicKey', key: string } }>, scripts?: Array<{ __typename?: 'WitnessScript', key: string, script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, bootstrap?: Array<{ __typename?: 'BootstrapWitness', signature: string, chainCode?: string | null | undefined, addressAttributes?: string | null | undefined, key?: { __typename?: 'PublicKey', hash: string, key: string } | null | undefined }> | null | undefined, redeemers?: Array<{ __typename?: 'Redeemer', index: number, purpose: string, scriptHash: string, executionUnits: { __typename?: 'ExecutionUnits', memory: number, steps: number } }> | null | undefined, datums?: Array<{ __typename?: 'Datum', hash: string, datum: string }> | null | undefined }, auxiliaryData?: { __typename?: 'AuxiliaryData', hash: string, body: { __typename?: 'AuxiliaryDataBody', scripts?: Array<{ __typename?: 'AuxiliaryScript', script: { __typename: 'NativeScript', any?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, all?: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> | null | undefined, nof?: Array<{ __typename?: 'NOf', key: string, scripts: Array<{ __typename: 'NativeScript', startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined }> }> | null | undefined, startsAt?: { __typename?: 'Slot', number: number } | null | undefined, expiresAt?: { __typename?: 'Slot', number: number } | null | undefined, vkey?: { __typename?: 'PublicKey', key: string } | null | undefined } | { __typename: 'PlutusScript', cborHex: string } }> | null | undefined, blob?: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray', array: Array<{ __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap', map: Array<{ __typename?: 'KeyValueMetadatum', key: string, metadatum: { __typename: 'BytesMetadatum', bytes: string } | { __typename: 'IntegerMetadatum', int: number } | { __typename: 'MetadatumArray' } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string }> } | { __typename: 'MetadatumMap' } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> } | { __typename: 'StringMetadatum', string: string } }> | null | undefined } } | null | undefined } }> } | null | undefined> | null | undefined };
 
 export const AllPoolParameterFieldsFragmentDoc = gql`
     fragment allPoolParameterFields on PoolParameters {
@@ -12215,6 +12215,10 @@ export const CoreTransactionFieldsFragmentDoc = gql`
     }
     hash
   }
+  requiredExtraSignatures {
+    hash
+  }
+  scriptIntegrityHash
   witness {
     signatures {
       publicKey {
@@ -12254,6 +12258,11 @@ export const CoreTransactionFieldsFragmentDoc = gql`
   auxiliaryData {
     hash
     body {
+      scripts {
+        script {
+          ...anyScript
+        }
+      }
       blob {
         key
         metadatum {
